requirements.txt
flask==2.1.3
Jinja2<3.1
cachelib
python-a2s
Werkzeug<2.1